Add `ignore_theme_config`
Although the theme does not include its `_config.yml` in the released gem, it gets merged with this site's config when the theme is downloaded from GitHub (as specified in the current `Gemfile`, and for any PR branch). That leads to unwanted settings, especially for `aux_links` and `compress_html`.

This setting should avoid this problem. It requires use of Jekyll 4.1.0 or later (https://jekyllrb.com/docs/configuration/options/).

See also https://github.com/just-the-docs/just-the-docs/pull/1244#issuecomment-1656867230.